refactor: move utils to prefecture dir
